add missing plates in data (#8)
Per #7, after the last pull request the pipeline would no longer run
because plates were added to `config.yml` but the actual `data`
subdirectories containing those plates were not added.

This pull request adds those missing plates from @ckikawa's directory
on the Hutch server running her local version of the analysis. The
results etc are not tracked yet (that needs to be done later), but
this minimally adds all the data so that the pipeline can now run
with its existing configuration and then adding the results can
be done in later pull requests.

@ckikawa, can you review and merge if you don't see any issues?