organize strains in viral libraries
Previously there were two sets of files in `./data/` with information about
the strains in the libraries; the CSVs in `data/viral_libraries/` and
FASTA files in `data/ha_sequences`, with the latter only containing
recent sequences.

Storing these data in two places with slightly different strain names
made it harder to keep track of data and do things like easily annotate
which barcodes correspond to which strain type (circulating 2025 vs vaccine).

This pull request removes `data/ha_sequences` altogether and just
makes all the information in `data/viral_libraries`. So now those
CSV files consolidate all this information in one place avoiding the
issue of different files with different information about the strain
libraries.

A rule was then added to the pipeline in `Snakefile` to create FASTA
files with the sequence data as **outputs** from the centralized CSV,
and these are placed in `results/viral_strain_seqs`.

Overall, this better organizes the viral library information used
by the pipeline in a single location.

@ckikawa, can you review and merge?