add `aggregate_human_sera_metadata_and_titers` (#23)
This custom rule in `Snakefile` makes aggregated CSVs of the sera
for which we have titers and the titers.

These CSVs are in `results/aggregated_analyses` and are tracked in the repo,
and described in the README.

Specification of which sera to include is configurable in `config.yml`,
this makes it configurable which sera to keep in these output files.