modularize custom rules to be configurable (#47)
This pull request does not actually change anything that affects the output or plots.
However, it does better modularize the parts of the pipeline that are custom rules
so that the code and pipeline can all be re-used for analyses in future years with
**only** changing the configuration in `config.yml` and not changing any of the code
or `Snakefile`. The issue was that before a few project specific variables (like
the *circulating_2025* strain type or library names that had *2025* in them) were
used in a way that ended up having them hardcoded in notebooks / scripts / `Snakefile`
in a way that would require changing that code if the next library is (for instance)
called *circulating_2025-2026* or something. This pull request now modularizes
everything by moving all of that configuration into `config.yml`.

@ckikawa, can you review and merge?